# Lesson2/templates/openapi.yaml
openapi: 3.0.3
info:
  title: Library API
  version: "1.0.0"
  description: API for a simple library system (books, users, borrows). Based on provided Flask endpoints.
servers:
  - url: http://localhost:3001
...
paths:
  /books/:
    get:
      tags:
        - Books
      summary: Find books
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Search string for title or author (optional). Also used as cache key.
      responses:
        '200':
          description: List of books found (from cache or DB)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '404':
          description: No books found
    post:
      tags:
        - Books
      summary: Add a new book
      description: Requires librarian role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - num_copies
              properties:
                title:
                  type: string
                author:
                  type: string
                num_copies:
                  type: integer
      responses:
        '201':
          description: Book added
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  id:
                    type: integer
        '401':
          description: Unauthorized (invalid/absent token or role)

      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        num_copies:
          type: integer
...